shmup: the 294 missing triangles were a loop bound and a float narrowing
Closed, both causes cited. `staticTris` now lands on the golden's 585630 exactly.

**The method is the reusable part.** Three witness numbers cannot localise a
geometry defect, so: a **per-emit trace** — the ordered list of every
`Assembler.add` with its palette key and triangle delta — taken from BOTH sides.
The original's comes from running `apps/shmup/src/world/*.js` **headless under
Node**: import the modules, seed an `Rng` with the world fork's state, stub
`materials.get`, run `index.js`'s init by hand. It reproduces the golden witness
exactly, so the trace is trustworthy. Both traces are **17,133 entries** and now
agree entry for entry. How to rebuild it is written into `scene/level.rs`'s
module doc so it survives the session.

**230 triangles — `ground.js:205`.** `for (let k = 0; k < sr.int(1, 3); k++)`
puts the bound in the loop CONDITION, so the source redraws it before every
iteration **including the one that fails** — one to four draws. The port hoisted
it to a single draw, desynchronising the seam pass's own fixed-seed stream
(`new Rng(0x5ea31d)`) from the first seam vertex: the `sr.float() < 0.22` skips
and the patch keys diverged. `road_dust` -130, `concrete` -170, `sand` +20,
`dirt` +40, `gravel` +10. The port already had `int_loop_continues` for this exact
idiom, used at **24 other sites**; this was the one that was not.

**64 triangles — `util.js:480`.** `Math.max(3, Math.round(w / 1.2))`. At the shop
facade's `w = 11.4`: f64 gives `9.500000000000002` -> **10**, f32 gives
`9.4999995` -> **9**. One jag vertex fewer is 8 triangles, on two panels each in
W1, W4, E1 and BS3. Measured, not inferred — the original's `wallPanel` fed those
exact parameters under Node returns 512/548 against the port's 504/540. The root
cause is the narrowing, not the division, so the width is now carried at the
source's precision to the one place that turns it into an integer.

**And a defect no count could ever have found.** Two `rng.pick` calls drawn AFTER
their transform instead of before (`interiors.js:503-508`, `:622-625`). JS
evaluates arguments left to right, so the colour comes off the stream first.
Either order consumes the same four values — so the witness passes either way —
it just hands each cushion the colour that belonged to its neighbour's x-offset.
Net **zero** triangles. Visible only in the frame.

    after                        staticTris    delta
    start                          585336       -294
    seam loop bound                585566        -64
    f64 division only              585614        -16
    f64 width (exact footprint)    585630          0

**The clutter verdict: they are road/ground DECALS, and the original draws every
one of them.** The policy is applied and correct — exactly nine prototypes survive
`ARENA_FLOOR` in both, with identical counts (`ac_unit` 109, `conduit_box` 58,
`water_tank` 45, `sat_dish` 43, `roof_vent` 32, `sign_hang` 8, `lamp_post` 5,
`lamp_glass` 5, `sign_board` 3 = **308**), and **7,771** ground placements are
discarded in both. Now pinned by a test.

What differs is how they read. The original lays them over a grey-green road
under 19 procedural surface textures, so `sand` against `road_dust` against
`concrete` is three materials. The port draws them flat-albedo against a warm
road, so those keys collapse to near-identical tan and their hard polygon edges
become the loudest thing in frame. That independently confirms the light-rig
attribution: baked asphalt is neutral 1.000:0.982:0.969 and the port renders
1.000:0.683:0.433.

(The dark "plank" mid-street in the port capture is the awning's SHADOW. The same
chevron is in the original, softer-edged.)

RNG neutrality: nothing was suppressed and no placement moved. The seam fix
touches only `sr`, which `seam` never receives — the source's own comment says
that stream exists so the seam scatter cannot shift the level's placement
sequence. The width fix draws nothing. The two pick reorders consume the same four
values at the same position. `instances`/`drawCalls` stay **308 / 62** throughout.

`the_static_triangle_count_still_falls_short` is now
`the_static_triangle_count_matches_the_golden`, un-ignored and passing. world::
227/227, integration 205 across 9 targets.

+++ apps/axiom-shmup/src/world/kit/facade.rs
@@ -14,7 +14,11 @@ use super::{solid_slabs, wall_panel, WallHole, WallPanelOpts, WallTop};
 /// arguments; the source's defaults are named here: `bevel = 0.022`,
 /// `top = Flat { jag: 0.0 }`, `warp = 0.018`.
 pub struct FacadeSpec<'a> {
-    pub w: f32,
+    /// The panel width at the SOURCE's f64 precision (`sideLen(spec, side)`,
+    /// `buildings.js:79`). [`wall_panel`] turns it into two integer step
+    /// counts, and an f32 width rounds one of them differently — see
+    /// `wall_panel`'s jag branch.
+    pub w: f64,
     pub h: f32,
     pub t: f32,
     pub key: &'a str,
@@ -66,7 +70,7 @@ pub fn facade_wall<'a>(asm: &mut Assembler, pm: &Mat4, spec: FacadeSpec<'a>) -> 
     asm.add_once(spec.key, &g, Some(pm), opts);
 
     let surface = asm.surface_of(spec.key);
-    for s in solid_slabs(spec.w, spec.h, spec.openings) {
+    for s in solid_slabs(spec.w as f32, spec.h, spec.openings) {
         asm.slab_box(surface, pm, s.x, s.y, s.w, s.h, spec.t);
     }
